test: update snapshots
