+++ packages/stats-validator-plugin-webpack/src/rules/diff-entry-download-size-limits/index.ts
@@ -210,7 +210,7 @@ const diffEntryDownloadSizeLimits: WebpackRule<Params> = (
     ]
   })
   .[entrypoints]`;
-  const result = data.query(query, data.files[0], {
+  const result = data.query(query, data.files, {
     params: normalizedParams,
   }) as ResultItem[];
