+++ packages/types/types/config/index.d.ts
@@ -1,10 +1,6 @@
 import { Config as ValidatorConfig } from '../validation/config';
-import { Report } from '../custom-report';
 
 export type Config = {
   silent?: boolean;
   validate?: ValidatorConfig;
-  ui?: {
-    customReports?: Report<unknown, unknown>[];
-  };
 };
