+++ test/bundles/v5/simple/webpack.config.js
@@ -27,6 +27,28 @@ module.exports = {
       statsOptions: {
         context,
       },
+      reports: [
+        {
+          id: 'top-20-biggest-modules',
+          name: 'Top 20 biggest modules',
+          data: { some: { custom: 'data' } }, // or () => fetchAsyncData()
+          view: [
+            'struct',
+            {
+              data: `#.stats.compilations.(
+            $compilation: $;
+            modules.({
+              module: $,
+              hash: $compilation.hash,
+              size: getModuleSize($compilation.hash)
+            })
+          ).sort(size.size desc)[:20]`,
+              view: 'list',
+              item: 'module-item',
+            },
+          ],
+        },
+      ],
     }),
   ],
   module: {
