+++ test/fixtures/cli/validate/config.js
@@ -8,29 +8,4 @@ module.exports = {
       'webpack/restricted-packages': ['error', ['foo']],
     },
   },
-
-  ui: {
-    customReports: [
-      {
-        id: 'top-20-biggest-modules',
-        name: 'Top 20 biggest modules',
-        data: { some: { custom: 'data' } },
-        view: [
-          'struct',
-          {
-            data: `#.stats.compilations.(
-              $compilation: $;
-              modules.({
-                module: $,
-                hash: $compilation.hash,
-                size: getModuleSize($compilation.hash)
-              })
-            ).sort(size.size desc)[:20]`,
-            view: 'list',
-            item: 'module-item',
-          },
-        ],
-      },
-    ],
-  },
 };
